Use ULL literals

+++ src/prime_generator.cpp
@@ -204,7 +204,7 @@ std::vector<BigInteger> SegmentedSieveOfEratosthenes(const BigInteger& n)
 {
     // TODO: This should scale to the system.
     // It's 16 GB in bytes.
-    const BigInteger limit = BigInteger(137438953472U);
+    const size_t limit = BigInteger(137438953472ULL);
 
     // `backward(n)` counts assuming that multiples
     // of 2 and 3 have been removed.
